kill some warnings

+++ src/bridges/PhysicalStratum_impl.cpp
@@ -32,14 +32,14 @@ void PhysicalStratum::forward() {
   // TODO: benchmark to see whether we want more sophisticated
   //       thread pool.
   vector<thread> threads;
-  for (int i = 0; i < executors.size(); i++) {
+  for (size_t i = 0; i < executors.size(); i++) {
     threads.push_back(thread(_forward, executors[i]));
   }
-  for (int i = 0; i < executors.size(); i++) {
+  for (size_t i = 0; i < executors.size(); i++) {
     threads[i].join();
   }
   report_forward_last_transfer.end();
-  for (int i = 0; i < executors.size(); i++) {
+  for (size_t i = 0; i < executors.size(); i++) {
     report_forward_last_transfer.aggregate_onlystat(executors[i]->report_forward_last_transfer);
   }
   report_forward_history.aggregate(report_forward_last_transfer);
@@ -48,14 +48,14 @@ void PhysicalStratum::forward() {
 void PhysicalStratum::backward() {
   report_backward_updateweight_last_transfer.reset();
   vector<thread> threads;
-  for (int i = 0; i < executors.size(); i++) {
+  for (size_t i = 0; i < executors.size(); i++) {
     threads.push_back(thread(_backward, executors[i]));
   }
-  for (int i = 0; i < executors.size(); i++) {
+  for (size_t i = 0; i < executors.size(); i++) {
     threads[i].join();
   }
   report_backward_updateweight_last_transfer.end();
-  for (int i = 0; i < executors.size(); i++) {
+  for (size_t i = 0; i < executors.size(); i++) {
     report_backward_updateweight_last_transfer.aggregate_onlystat(executors[i]->report_backward_updateweight_last_transfer);
   }
   report_backward_updateweight_history.aggregate(report_backward_updateweight_last_transfer);
